requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.10
Flask-Migrate==4.0.5
marshmallow-sqlalchemy==0.29.0
Flask-RESTful==0.3.10
python-dotenv==1.0.0
Werkzeug==2.3.7
SQLAlchemy==2.0.39